style: break out desktop and mobile headers into components
